requirements.txt
aniso8601==9.0.1
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Flask==3.1.0
Flask-RESTful==0.3.10
flask-swagger==0.2.14
fonttools==4.55.0
future==1.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
netmiko==4.4.0
networkx==3.4.2
ntc_templates==7.4.0
numpy==2.1.3
packaging==24.2
paramiko==3.5.0
pillow==11.0.0
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.2.0
pyserial==3.5
pysnmp==7.1.13
pytest==8.3.3
python-dateutil==2.9.0.post0
python-nmap==0.7.1
pytz==2024.2
PyYAML==6.0.2
scp==0.15.0
setuptools==75.6.0
six==1.16.0
textfsm==1.1.3
Werkzeug==3.1.3
